--- cmd_install.go
@@ -1,41 +1,0 @@
-//
-// Copyright © 2021 Kris Nóva <kris@nivenly.com>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-//    ███╗   ██╗ ██████╗ ██╗   ██╗ █████╗
-//    ████╗  ██║██╔═████╗██║   ██║██╔══██╗
-//    ██╔██╗ ██║██║██╔██║██║   ██║███████║
-//    ██║╚██╗██║████╔╝██║╚██╗ ██╔╝██╔══██║
-//    ██║ ╚████║╚██████╔╝ ╚████╔╝ ██║  ██║
-//    ╚═╝  ╚═══╝ ╚═════╝   ╚═══╝  ╚═╝  ╚═╝
-
-package naml
-
-import (
-	"github.com/kris-nova/logger"
-)
-
-// Install is used to install an application in Kubernetes
-func Install(app Deployable) error {
-	client, err := Client()
-	if err != nil {
-		return err
-	}
-	err = app.Install(client)
-	if err != nil {
-		return err
-	}
-	logger.Success("Successfully installed [%s]", app.Meta().Name)
-	return nil
-}
