+++ go.mod
@@ -5,7 +5,6 @@ go 1.16
 require (
 	github.com/kris-nova/logger v0.2.2
 	github.com/urfave/cli/v2 v2.3.0
-	k8s.io/api v0.21.1
 	k8s.io/apimachinery v0.21.1
 	k8s.io/client-go v0.21.1
 	sigs.k8s.io/kind v0.11.1
